fix: Quality layer flag/mask can zero, override or pass-through previous layers
- quality layer flag/bitmask values can be NaN - this allows earlier layer value to flow through
- means that days with no quality layer data just inherit the lower layers values
- can also use zero to clear flags/bitmasks

+++ src/mag_toolkit/calibration/CalibrationLayer.py
@@ -149,13 +149,13 @@ class CalibrationLayer(Layer):
         timedelta_var.attrs["DEPEND_0"] = CONSTANTS.CDF_VARS.EPOCH
         qf_var = xr.Variable(
             dims=[CONSTANTS.CDF_VARS.EPOCH],
-            data=self._contents[CONSTANTS.CSV_VARS.QUALITY_FLAG],
+            data=self._contents[CONSTANTS.CSV_VARS.QUALITY_FLAG].astype(int),
             attrs=skeleton_cdf[CONSTANTS.CDF_VARS.QUALITY_FLAG].attrs,
         )
         qf_var.attrs["DEPEND_0"] = CONSTANTS.CDF_VARS.EPOCH
         qb_var = xr.Variable(
             dims=[CONSTANTS.CDF_VARS.EPOCH],
-            data=self._contents[CONSTANTS.CSV_VARS.QUALITY_BITMASK],
+            data=self._contents[CONSTANTS.CSV_VARS.QUALITY_BITMASK].astype(int),
             attrs=skeleton_cdf[CONSTANTS.CDF_VARS.QUALITY_BITMASK].attrs,
         )
         qb_var.attrs["DEPEND_0"] = CONSTANTS.CDF_VARS.EPOCH
